requirements.txt
django==2.2.10
djangorestframework==3.12.2
django-environ==0.4.5
psycopg2==2.8.6
drf-yasg
